+++ R/weighting.R
@@ -47,8 +47,9 @@ function (x, weights = NULL, normwt = "ignored", na.rm = TRUE)
 #' @param y another optional vector for a two-way frequency table. Must be the same length as \code{x}
 #' @param weights vector of weights, must be the same length as \code{x}
 #' @param normwt if TRUE, normalize weights so that the total weighted count is the same as the unweighted one
-#' @param na.show if TRUE, show NA count in table output
-#' @param na.rm if TRUE, remove NA values before computation
+#' @param useNA wether to include NA values in the table
+#' @param na.show (deprecated) if TRUE, show NA count in table output
+#' @param na.rm (deprecated) if TRUE, remove NA values before computation
 #' @param digits Number of significant digits.
 #' @param exclude values to remove from x and y. To exclude NA, use na.rm argument.
 #' @details
@@ -70,16 +71,26 @@ function (x, weights = NULL, normwt = "ignored", na.rm = TRUE)
 
 
 `wtd.table` <-
-function (x, y = NULL, weights = NULL, digits = 3, normwt = FALSE, na.rm = TRUE, na.show = FALSE, exclude = NULL) 
+function (x, y = NULL, weights = NULL, digits = 3, normwt = FALSE, useNA = c("no", "ifany", "always"), na.rm = TRUE, na.show = FALSE, exclude = NULL) 
 {
   if (is.null(weights)) weights <- rep(1, length(x))  
   if (length(x) != length(weights)) stop("x and weights lengths must be the same")
   if (!is.null(y) & (length(x) != length(y))) stop("x and y lengths must be the same")
-  if (na.show) {
-      x <- addNA(x)
-      if (!is.null(y)) y <- addNA(y)
+  miss.usena <- missing(useNA)
+  useNA <- match.arg(useNA)
+
+  if (!missing(na.show) || !missing(na.rm)) {
+    if (miss.usena) warning("'na.rm' and 'na.show' are deprecated. Use 'useNA' instead.")
+    else warning("'na.rm' and 'na.show' are ignored when 'useNA' is provided.")
   }
-  if (na.rm) {
+  if (useNA != "no" || (na.show && miss.usena)) {
+    if (match(NA, exclude, nomatch = 0L)) {
+      warning("'exclude' containing NA and 'useNA' != \"no\"' are a bit contradicting")
+    }
+    x <- addNA(x)
+    if (!is.null(y)) y <- addNA(y)
+  }
+  if (useNA == "no" || (na.rm && miss.usena)) {
      s <- !is.na(x) & !is.na(weights)
      if (!is.null(y)) s <- s & !is.na(y)
      x <- x[s, drop = FALSE]
@@ -103,7 +114,12 @@ function (x, y = NULL, weights = NULL, digits = 3, normwt = FALSE, na.rm = TRUE,
     result <- tapply(weights, list(x,y), sum, simplify=TRUE)
   }
   result[is.na(result)] <- 0
-  as.table(result)
+  tab <- as.table(result)
+  if (useNA == "ifany") {
+    if (sum(tab[,is.na(colnames(tab))]) == 0) tab <- tab[,!is.na(colnames(tab))]
+    if (sum(tab[is.na(rownames(tab)),]) == 0) tab <- tab[!is.na(rownames(tab)),]
+  }
+  tab
 }
 
 ##' @export